% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unjoin.r
\name{unjoin}
\alias{unjoin}
\alias{unjoin}
\alias{unjoin.unjoin}
\alias{unjoin}
\alias{unjoin.data.frame}
\title{Split a table in two and remove repeated values.}
\usage{
unjoin(data, ...)

\method{unjoin}{unjoin}(data, ..., key_col = ".idx0")

\method{unjoin}{data.frame}(data, ..., key_col = ".idx0")
}
\arguments{
\item{data}{A data frame.}

\item{...}{Specification of columns to nest. Use bare variable names.
Select all variables between x and z with \code{x:z}, exclude y with
\code{-y}. For more options, see the \link[dplyr]{select} documentation.}

\item{key_col}{The name of the new column to key the two output data frames.}
}
\description{
This is a subset of the tasks done by \code{\link[tidyr]{nest}}.
}
\examples{
library(dplyr)
data("Seatbelts", package= "datasets")
x <- unjoin(as.data.frame(Seatbelts), front, rear, kms)
all.equal(inner_join(x$.idx0, x$data) \%>\% select(-.idx0), as.data.frame(Seatbelts))

iris \%>\% unjoin(-Species)
chickwts \%>\% unjoin(weight)

if (require("gapminder")) {
  gapminder \%>\%
    group_by(country, continent) \%>\%
    unjoin()

  gapminder \%>\%
    unjoin(-country, -continent)
  unjoin(gapminder)
}
unjoin(iris, Petal.Width)
}
\seealso{
`dplyr::inner_join` for the inverse operation.

`tidyr::nest` for the complementary operation resulting in one data frame
}
