% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unjoin.r
\name{unjoin_}
\alias{unjoin_}
\title{Standard-evaluation version of 'unjoin'.}
\usage{
unjoin_(data, unjoin_cols = character(), key_col = ".idx0")
}
\arguments{
\item{data}{A data frame.}

\item{unjoin_cols}{Character vector of columns to nest.}

\item{key_col}{The name of the new column to key the two output data frames.}
}
\description{
This is a S3 generic.
}
\keyword{internal}
